# apps/backend-template/src/modules/Users/adapters/in/http/controllers/OrganizationController.ts
import { INFRA_NOT_IMPLEMENTED } from '@src/config/constants';
import { ForbiddenError } from '@src/infra/exceptions';
import Security from '@src/infra/security';
import BaseController from '@src/interface/HTTP/ports/BaseController';
import { validateRequestAgainstOAS } from '@src/interface/HTTP/validators';
import { setListQuery, setMetricsQuery } from '@src/modules/port';
import BaseDomainEvent from '@src/modules/port/BaseDomainEvent';
import {
  decideOrganizationAccess,
  resolveOrganizationCollectionScope
} from '@src/modules/Users/domain/security/TenantAuthorizationPolicy';
import Authorize from '@src/shared/decorators/guard/Authorize';

import type { IController, IControllerFactory } from '@src/interface/HTTP/ports';
import type { IServiceResponse } from '@src/modules/port';
import type { IOrganizationUseCases } from '@src/modules/Users/application/ports/IOrganizationUseCases';
import type { IOrganization } from '@src/modules/Users/domain/Entity/IOrganization';
import type { ITenantAuthorizationDecision } from '@src/modules/Users/domain/security/TenantAuthorizationPolicy';
import type { RequestCreateAddress } from '@src/modules/Users/interface/dto/RequestCreateAddress';
import type { RequestCreateEmail } from '@src/modules/Users/interface/dto/RequestCreateEmail';
import type { RequestCreateOrganization } from '@src/modules/Users/interface/dto/RequestCreateOrganization';
import type { RequestCreatePhone } from '@src/modules/Users/interface/dto/RequestCreatePhone';
import type { RequestUpdateAddress } from '@src/modules/Users/interface/dto/RequestUpdateAddress';
import type { RequestUpdateEmail } from '@src/modules/Users/interface/dto/RequestUpdateEmail';
import type { RequestUpdateOrganization } from '@src/modules/Users/interface/dto/RequestUpdateOrganization';
import type { RequestUpdatePhone } from '@src/modules/Users/interface/dto/RequestUpdatePhone';

class OrganizationController extends BaseController implements IController {
  private readonly organizationUseCases: IOrganizationUseCases;

  constructor(factory: IControllerFactory) {
    super(factory);
    if (!factory.organizationUseCases) {
      const error = new Error('OrganizationUseCases is not implemented');
      error.name = INFRA_NOT_IMPLEMENTED;
      throw error;
    }
    this.organizationUseCases = factory.organizationUseCases;
  }

  // eslint-disable-next-line class-methods-use-this
  private getAuthenticatedUser(event: BaseDomainEvent): Record<string, any> {
    return ((event as any).authenticatedUser || {}) as Record<string, any>;
  }

  // eslint-disable-next-line class-methods-use-this
  private throwIfTenantAccessDenied(decision: ITenantAuthorizationDecision): void {
    if (!decision.allowed) {
      throw new ForbiddenError(decision.reason);
    }
  }

  private enforceOrganizationScope(event: BaseDomainEvent, organizationId: string): void {
    const authenticatedUser = this.getAuthenticatedUser(event);
    this.throwIfTenantAccessDenied(decideOrganizationAccess(authenticatedUser, organizationId));
  }

  @Authorize()
  public async create(event: BaseDomainEvent): Promise<IServiceResponse<IOrganization>> {
    validateRequestAgainstOAS(this.openApiSpecification, event.schemaOAS, event);
    const requestCreateOrganization = event.input as RequestCreateOrganization;
    const { result, error } = await this.organizationUseCases.create(requestCreateOrganization);
    return { result, error };
  }

  @Authorize()
  public async update(event: BaseDomainEvent): Promise<IServiceResponse<IOrganization>> {
    validateRequestAgainstOAS(this.openApiSpecification, event.schemaOAS, event);
    const requestUpdateOrganization = event.input as RequestUpdateOrganization;
    const organizationId = Security.xss(event.params.id);
    this.enforceOrganizationScope(event, organizationId);
    const { result, error } = await this.organizationUseCases.update(
      organizationId,
      requestUpdateOrganization
    );
    return { result, error };
  }

  @Authorize()
  public async delete(event: BaseDomainEvent): Promise<IServiceResponse<boolean>> {
    validateRequestAgainstOAS(this.openApiSpecification, event.schemaOAS, event);
    const organizationId = Security.xss(event.params.id);
    this.enforceOrganizationScope(event, organizationId);
    const { result, error } = await this.organizationUseCases.delete(organizationId);
    return { result, error };
  }

  @Authorize()
  public async getOneById(event: BaseDomainEvent): Promise<IServiceResponse<IOrganization>> {
    validateRequestAgainstOAS(this.openApiSpecification, event.schemaOAS, event);
    const organizationId = Security.xss(event.params.id);
    this.enforceOrganizationScope(event, organizationId);
    const { result, error } = await this.organizationUseCases.getOneById(organizationId);
    return { result, error };
  }

  @Authorize()
  public async getAll(event: BaseDomainEvent): Promise<IServiceResponse<IOrganization[]>> {
    validateRequestAgainstOAS(this.openApiSpecification, event.schemaOAS, event);
    // JUM-777: page/size/filter/sort/q validated against x-list-capabilities.
    const { filters, paging } = setListQuery(event);
    const authenticatedUser = this.getAuthenticatedUser(event);
    const tenantScope = resolveOrganizationCollectionScope(authenticatedUser);
    this.throwIfTenantAccessDenied(tenantScope.decision);
    return this.organizationUseCases.getAll({ ...filters, ...tenantScope.filters }, paging);
  }

  @Authorize()
  public async getOrganizationsMetrics(event: BaseDomainEvent): Promise<IServiceResponse<unknown>> {
    validateRequestAgainstOAS(this.openApiSpecification, event.schemaOAS, event);
    const { query, filters, capabilities } = setMetricsQuery(event);
    const authenticatedUser = this.getAuthenticatedUser(event);
    const tenantScope = resolveOrganizationCollectionScope(authenticatedUser);
    this.throwIfTenantAccessDenied(tenantScope.decision);
    return this.organizationUseCases.metrics(
      { ...filters, ...tenantScope.filters },
      query,
      capabilities
    );
  }

  @Authorize()
  public async createAddress(event: BaseDomainEvent): Promise<IServiceResponse<IOrganization>> {
    validateRequestAgainstOAS(this.openApiSpecification, event.schemaOAS, event);
    const organizationId = Security.xss(event.params.id);
    this.enforceOrganizationScope(event, organizationId);
    const requestCreateAddress = event.input as RequestCreateAddress;
    const { result, error } = await this.organizationUseCases.createAddress(
      organizationId,
      requestCreateAddress
    );
    return { result, error };
  }

  @Authorize()
  public async updateAddress(event: BaseDomainEvent): Promise<IServiceResponse<IOrganization>> {
    validateRequestAgainstOAS(this.openApiSpecification, event.schemaOAS, event);
    const organizationId = Security.xss(event.params.id);
    this.enforceOrganizationScope(event, organizationId);
    const addressId = Security.xss(event.params.addressId);
    const requestUpdateAddress = event.input as RequestUpdateAddress;
    const { result, error } = await this.organizationUseCases.updateAddress(
      organizationId,
      addressId,
      requestUpdateAddress
    );
    return { result, error };
  }

  @Authorize()
  public async deleteAddress(event: BaseDomainEvent): Promise<IServiceResponse<IOrganization>> {
    validateRequestAgainstOAS(this.openApiSpecification, event.schemaOAS, event);
    const organizationId = Security.xss(event.params.id);
    this.enforceOrganizationScope(event, organizationId);
    const addressId = Security.xss(event.params.addressId);
    const { result, error } = await this.organizationUseCases.deleteAddress(
      organizationId,
      addressId
    );
    return { result, error };
  }

  @Authorize()
  public async createPhone(event: BaseDomainEvent): Promise<IServiceResponse<IOrganization>> {
    validateRequestAgainstOAS(this.openApiSpecification, event.schemaOAS, event);
    const organizationId = Security.xss(event.params.id);
    this.enforceOrganizationScope(event, organizationId);
    const requestCreatePhone = event.input as RequestCreatePhone;
    const { result, error } = await this.organizationUseCases.createPhone(
      organizationId,
      requestCreatePhone
    );
    return { result, error };
  }

  @Authorize()
  public async updatePhone(event: BaseDomainEvent): Promise<IServiceResponse<IOrganization>> {
    validateRequestAgainstOAS(this.openApiSpecification, event.schemaOAS, event);
    const organizationId = Security.xss(event.params.id);
    this.enforceOrganizationScope(event, organizationId);
    const phoneId = Security.xss(event.params.phoneId);
    const requestUpdatePhone = event.input as RequestUpdatePhone;
    const { result, error } = await this.organizationUseCases.updatePhone(
      organizationId,
      phoneId,
      requestUpdatePhone
    );
    return { result, error };
  }

  @Authorize()
  public async deletePhone(event: BaseDomainEvent): Promise<IServiceResponse<IOrganization>> {
    validateRequestAgainstOAS(this.openApiSpecification, event.schemaOAS, event);
    const organizationId = Security.xss(event.params.id);
    this.enforceOrganizationScope(event, organizationId);
    const phoneId = Security.xss(event.params.phoneId);
    const { result, error } = await this.organizationUseCases.deletePhone(organizationId, phoneId);
    return { result, error };
  }

  @Authorize()
  public async createEmail(event: BaseDomainEvent): Promise<IServiceResponse<IOrganization>> {
    validateRequestAgainstOAS(this.openApiSpecification, event.schemaOAS, event);
    const organizationId = Security.xss(event.params.id);
    this.enforceOrganizationScope(event, organizationId);
    const requestCreateEmail = event.input as RequestCreateEmail;
    const { result, error } = await this.organizationUseCases.createEmail(
      organizationId,
      requestCreateEmail
    );
    return { result, error };
  }

  @Authorize()
  public async updateEmail(event: BaseDomainEvent): Promise<IServiceResponse<IOrganization>> {
    validateRequestAgainstOAS(this.openApiSpecification, event.schemaOAS, event);
    const organizationId = Security.xss(event.params.id);
    this.enforceOrganizationScope(event, organizationId);
    const emailId = Security.xss(event.params.emailId);
    const requestUpdateEmail = event.input as RequestUpdateEmail;
    const { result, error } = await this.organizationUseCases.updateEmail(
      organizationId,
      emailId,
      requestUpdateEmail
    );
    return { result, error };
  }

  @Authorize()
  public async deleteEmail(event: BaseDomainEvent): Promise<IServiceResponse<IOrganization>> {
    validateRequestAgainstOAS(this.openApiSpecification, event.schemaOAS, event);
    const organizationId = Security.xss(event.params.id);
    this.enforceOrganizationScope(event, organizationId);
    const emailId = Security.xss(event.params.emailId);
    const { result, error } = await this.organizationUseCases.deleteEmail(organizationId, emailId);
    return { result, error };
  }

  // Backward-compatible aliases for existing operationIds/handlers.
  public createOrganization(event: BaseDomainEvent): Promise<IServiceResponse<IOrganization>> {
    return this.create(event);
  }

  public updateOrganization(event: BaseDomainEvent): Promise<IServiceResponse<IOrganization>> {
    return this.update(event);
  }

  public deleteOrganization(event: BaseDomainEvent): Promise<IServiceResponse<boolean>> {
    return this.delete(event);
  }

  public getOrganizationById(event: BaseDomainEvent): Promise<IServiceResponse<IOrganization>> {
    return this.getOneById(event);
  }

  public getAllOrganizations(event: BaseDomainEvent): Promise<IServiceResponse<IOrganization[]>> {
    return this.getAll(event);
  }

  public createOrganizationAddress(
    event: BaseDomainEvent
  ): Promise<IServiceResponse<IOrganization>> {
    return this.createAddress(event);
  }

  public updateOrganizationAddress(
    event: BaseDomainEvent
  ): Promise<IServiceResponse<IOrganization>> {
    return this.updateAddress(event);
  }

  public deleteOrganizationAddress(
    event: BaseDomainEvent
  ): Promise<IServiceResponse<IOrganization>> {
    return this.deleteAddress(event);
  }

  public createOrganizationPhone(event: BaseDomainEvent): Promise<IServiceResponse<IOrganization>> {
    return this.createPhone(event);
  }

  public updateOrganizationPhone(event: BaseDomainEvent): Promise<IServiceResponse<IOrganization>> {
    return this.updatePhone(event);
  }

  public deleteOrganizationPhone(event: BaseDomainEvent): Promise<IServiceResponse<IOrganization>> {
    return this.deletePhone(event);
  }

  public createOrganizationEmail(event: BaseDomainEvent): Promise<IServiceResponse<IOrganization>> {
    return this.createEmail(event);
  }

  public updateOrganizationEmail(event: BaseDomainEvent): Promise<IServiceResponse<IOrganization>> {
    return this.updateEmail(event);
  }

  public deleteOrganizationEmail(event: BaseDomainEvent): Promise<IServiceResponse<IOrganization>> {
    return this.deleteEmail(event);
  }

  public static compile(factory: IControllerFactory) {
    return new OrganizationController(factory);
  }
}

export default OrganizationController;