# packages/cli-init/templates/frontend/src/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jumentix API",
    "description": "`REST API` built with `Jumentix` and `Domain Designer`",
    "termsOfService": "https://www.xpertminds.dev/terms/",
    "contact": {
      "email": "apiteam@xpertminds.dev"
    },
    "license": {
      "name": "Unlicensed",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0",
    "x-rbac": {
      "description": "Role-to-scope matrix enforced by the backend and mirrored here for OAS-driven clients.",
      "roles": {
        "superadmin": ["*"],
        "admin": [
          "access_allow",
          "read_user",
          "create_user",
          "update_user",
          "delete_user",
          "read_organization",
          "update_organization"
        ],
        "user": ["access_allow", "read_user"]
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/1.0.0",
      "x-service-id": "core"
    }
  ],
...
  "paths": {
...
    "/users/{id}": {
      "delete": {
        "tags": ["user"],
        "summary": "Delete user",
        "description": "Deletes a single user",
        "operationId": "deleteOne",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceDeleteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearerAuth": ["delete_user"]
          }
        ]
      },
      "put": {
        "tags": ["user"],
        "summary": "Update user",
        "description": "Updates a single user",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Update an user from Store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestUpdateUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "security": [
          {
            "bearerAuth": ["update_user"]
          }
        ]
      },
      "get": {
        "tags": ["user"],
        "summary": "Get an user by ID",
        "description": "Returns a single user",
        "operationId": "getOneById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearerAuth": ["read_user"]
          }
        ]
      }
    },
...
    "/users/{id}/createEmail": {
      "post": {
        "tags": ["user"],
        "summary": "Create email address for a given user",
        "description": "Create email address for a given user",
        "operationId": "createEmail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Email create data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCreateEmail"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Email"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "security": [
          {
            "bearerAuth": ["update_user"]
          }
        ]
      }
    },
    "/users/{id}/updateEmail/{emailId}": {
      "put": {
        "tags": ["user"],
        "summary": "Update email address for a given user",
        "description": "Update email address for a given user",
        "operationId": "updateEmail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "emailId",
            "in": "path",
            "description": "emailId to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Email update data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestUpdateEmail"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Email"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "security": [
          {
            "bearerAuth": ["update_user"]
          }
        ]
      }
    },
    "/users/{id}/deleteEmail/{emailId}": {
      "delete": {
        "tags": ["user"],
        "summary": "Delete email address of an user",
        "description": "Deletes email address of an user",
        "operationId": "deleteEmail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "emailId",
            "in": "path",
            "description": "emailId to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceDeleteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Resource not found"
          }
        },
        "security": [
          {
            "bearerAuth": ["delete_user"]
          }
        ]
      }
    },
...
    "/users/{id}/createPhone": {
      "post": {
        "tags": ["user"],
        "summary": "Create phone address for a given user",
        "description": "Create phone address for a given user",
        "operationId": "createPhone",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Phone create data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCreatePhone"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phone"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "security": [
          {
            "bearerAuth": ["update_user"]
          }
        ]
      }
    },
    "/users/{id}/updatePhone/{phoneId}": {
      "put": {
        "tags": ["user"],
        "summary": "Update phone address for a given user",
        "description": "Update phone address for a given user",
        "operationId": "updatePhone",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phoneId",
            "in": "path",
            "description": "phoneId to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Phone update data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestUpdatePhone"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phone"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "security": [
          {
            "bearerAuth": ["update_user"]
          }
        ]
      }
    },
    "/users/{id}/deletePhone/{phoneId}": {
      "delete": {
        "tags": ["user"],
        "summary": "Delete phone of an user",
        "description": "Deletes phone of an user",
        "operationId": "deletePhone",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phoneId",
            "in": "path",
            "description": "phoneId to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceDeleteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Resource not found"
          }
        },
        "security": [
          {
            "bearerAuth": ["delete_user"]
          }
        ]
      }
    },
    "/organizations": {
      "get": {
        "tags": ["organization"],
        "summary": "Get all organizations",
        "description": "Returns a page of organizations. Delta sync page: sort=`updatedAt:asc,id:asc`, filter `{\"updatedAt\":{\"operator\":\"gt\",\"value\":\"<iso>\"}}`, page through `page`/`size`, set `includeDeleted=true` so tombstones travel with the cursor (`x-sync`).",
        "operationId": "getAllOrganizations",
        "x-service": "core",
        "x-sync": {
          "cursorField": "updatedAt",
          "includeDeletedParam": "includeDeleted"
        },
        "x-list-capabilities": {
          "sortable": ["name", "createdAt", "updatedAt", "id"],
          "filterable": {
            "name": "text",
            "createdAt": "date",
            "updatedAt": "date"
          },
          "searchable": ["name"],
          "defaultSize": 30,
          "maxSize": 100,
          "includeDeleted": true
        },
        "x-metrics-capabilities": {
          "groupable": [],
          "series": ["createdAt", "updatedAt"]
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number to fetch (1-based).",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Page size; bounded by `x-list-capabilities.maxSize`.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 30
            }
          },
          {
            "name": "filter",
            "in": "query",
            "description": "Base64-encoded JSON object keyed by a field listed in `x-list-capabilities.filterable`. A scalar value means equality; an object `{ \"operator\": \"contains\" | \"eq\" | \"ne\" | \"in\" | \"gte\" | \"lte\" | \"between\", \"value\": ... }` applies that operator. Unknown fields or operators are rejected with 400.",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 4096
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Comma-separated `field:asc|desc` pairs, fields limited to `x-list-capabilities.sortable`. Unknown fields are rejected with 400.",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 256,
              "pattern": "^[A-Za-z0-9_.]+(:(asc|desc))?(,[A-Za-z0-9_.]+(:(asc|desc))?)*$"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Case-insensitive free-text search over `x-list-capabilities.searchable`.",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "includeDeleted",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationArrayOf"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearerAuth": ["read_organization"]
          }
        ]
      },
      "post": {
        "tags": ["organization"],
        "summary": "Add a new organization",
        "description": "Add a new organization",
        "operationId": "createOrganization",
        "requestBody": {
          "description": "Create organization payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCreateOrganization"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Organization created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "security": [
          {
            "bearerAuth": ["create_organization"]
          }
        ]
      }
    },
...
    "/organizations/{id}/createAddress": {
      "post": {
        "tags": ["organization"],
        "summary": "Create organization address entry",
        "description": "Creates an address value object entry for an organization",
        "operationId": "createOrganizationAddress",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Address create payload",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCreateAddress"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Organization not found"
          }
        },
        "security": [
          {
            "bearerAuth": ["update_organization"]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ResourceDeleteResponse": {
        "description": "Port output object for delete operations.",
        "required": ["data"],
        "type": "object",
        "properties": {
          "data": {
            "x-label": {
              "en": "Number",
              "pt-BR": "Número"
            },
            "type": "boolean",
            "default": false,
            "description": "Result of request to delete resource"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}